feat(ehr): moving internal
Ref: #1040
