+++ packages/api/src/routes/index.ts
@@ -7,8 +7,8 @@ import ehr from "./ehr";
 import ehrInternal from "./ehr-internal";
 import feedback from "./feedback";
 import { reportClientErrors } from "./helpers/report-client-errors";
-import internal from "./internal";
-import jwtToken from "./jwt-token";
+import internal from "./internal/internal";
+import jwtToken from "./internal/jwt-token/ehr";
 import medical from "./medical";
 import fhirRouter from "./medical/fhir-r4-proxy";
 import { checkMAPIAccess, processCxId } from "./middlewares/auth";
